flask_serverless_copy/requirements.txt
aniso8601==7.0.0
anyio==3.6.1
ariadne==0.15.1
attrs==21.4.0
boto3==1.24.44
botocore==1.27.44
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
coverage==6.4.2
Flask==2.1.3
Flask-GraphQL==2.0.0
Flask-GraphQL-Auth==1.3.3
Flask-MySQLdb==1.0.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
graphene==2.1.9
graphene-pynamodb==2.3.0
graphene-sqlalchemy==2.3.0
graphql-core==2.3
graphql-relay==2.0.1
graphql-server-core==1.1.1
graphql-utils==0.4
greenlet==1.1.2
idna==3.3
importlib-metadata==4.12.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
mysqlclient==2.1.1
packaging==21.3
pluggy==1.0.0
promise==2.3
py==1.11.0
PyJWT==2.0.1
pynamodb==5.0.0
pyparsing==3.0.9
pytest==7.1.2
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
Rx==1.6.1
s3transfer==0.6.0
singledispatch==3.7.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.39
starlette==0.20.4
tomli==2.0.1
typing_extensions==4.3.0
urllib3==1.26.11
Werkzeug==2.2.0
wrapt==1.14.1
zipp==3.8.1
